First mini-Docker version

+++ internal/database/taskFunctions.go
@@ -9,7 +9,7 @@ import (
 	"github.com/jackc/pgx/v5"
 )
 
-func (db *Database) AddTask(title, description string, timeLimit int, memoryLimit int, isPractice bool) (int, error) {
+func (db *Database) AddTask(title, description string, timeLimit float64, memoryLimit float64, isPractice bool) (int, error) {
 	psql := sq.StatementBuilder.PlaceholderFormat(sq.Dollar)
 	sql, args, err := psql.Insert("tasks").Columns("title", "description", "time_limit", "memory_limit", "is_practice").Values(title, description, timeLimit, memoryLimit, isPractice).Suffix("RETURNING id").ToSql()
 	if err != nil {
@@ -40,7 +40,7 @@ func (db *Database) DeleteTask(id int) error {
 	return nil
 }
 
-func (db *Database) EditTask(id int, title, description string, timeLimit, memoryLimit int, isPractice bool) error {
+func (db *Database) EditTask(id int, title, description string, timeLimit, memoryLimit float64, isPractice bool) error {
 	psql := sq.StatementBuilder.PlaceholderFormat(sq.Dollar)
 	sql, args, err := psql.Update("tasks").Set("title", title).Set("description", description).Set("time_limit", timeLimit).Set("memory_limit", memoryLimit).Set("is_practice", isPractice).Where(sq.Eq{"id": id}).ToSql()
 	if err != nil {
